Finished admin page. Admin can create, edit, and delete new users

+++ src/main/java/com/capstone/capstoneproject/models/data/UserDao.java
@@ -5,6 +5,6 @@ import org.springframework.data.repository.CrudRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface UserDao extends CrudRepository<User, Long> {
+public interface UserDao extends CrudRepository<User, Integer> {
     User findByEmail(String email);
 }
